+++ CHANGELOG.md
@@ -9,11 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Parsing of `timeunit` and `timeprecision` statements. ([#92](https://github.com/fabianschuiki/moore/issues/92))
 - Parsing of `inside` expressions. ([#93](https://github.com/fabianschuiki/moore/issues/93))
 - Support for the `**` operators. ([#94](https://github.com/fabianschuiki/moore/issues/94))
-- Parse the `::` operator.
+- Support for the `::` operator.
 - Parse cast expressions.
 - Accept ``` `` ```, `` `\ ``, and `` `" `` in preprocessor.
 - Parse optional lifetime specifiers after `task` or `function`.
 - Support for enum types. ([#79](https://github.com/fabianschuiki/moore/issues/79))
+- Support for package declarations.
 
 ### Fixed
 - Continuous assignments resolve in epsilon time. ([#89](https://github.com/fabianschuiki/moore/issues/89))